backend/requirements.txt
fastapi[standard]
uvicorn[standard]
sqlalchemy
pymysql
bcrypt
PyJWT
python-dotenv
